Google Foundation of Cybersecurity
This course taught me a lot of things that I did not know. I got the chance to listen to cybersecurity professional and answer a lot of exciting questions. 
some of the main things I learned from this course was 1 Recognizing core skills and knowledge needed to become a cybersecurity analyst.
2. Identifying how security attacks impact business operations.
3. Explaining security ethics.
4. Identifying common tools used by cybersecurity analysts.